Fix issue causing errors on different line endings
Dipper was looking for PHP_EOL line-endings, but this wasn't universal
enough of a find. It was matching within your environment, but if you
were dealing with a file from another environment, your PHP_EOL and
the actual EOL set being used might be different. This adjustment now
adjusts hunts through for a line-ending and adjusts from there.

Also note, this is using a strpos() check followed by an explode()
because that's a faster two-step process than one preg_split() was in
my testing.